Chore: Cnpj file renamed to Cnpjs so it follows the pattern of models
